created new test file for new component
